+++ ui/actions/resources/index.ts
@@ -3,6 +3,7 @@ export {
   getLatestResources,
   getMetadataInfo,
   getResourceById,
+  getResourceDrawerData,
   getResourceEvents,
   getResources,
 } from "./resources";